split out inst rep tests
