+++ qcodes/tests/parameter/test_instrument_ref_parameter.py
@@ -1,0 +1,25 @@
+from unittest import TestCase
+
+from qcodes.instrument.parameter import InstrumentRefParameter
+from qcodes.tests.instrument_mocks import DummyInstrument
+
+
+class TestInstrumentRefParameter(TestCase):
+
+    def setUp(self):
+        self.a = DummyInstrument('dummy_holder')
+        self.d = DummyInstrument('dummy')
+
+    def test_get_instr(self):
+        self.a.add_parameter('test', parameter_class=InstrumentRefParameter)
+
+        self.a.test.set(self.d.name)
+
+        self.assertEqual(self.a.test.get(), self.d.name)
+        self.assertEqual(self.a.test.get_instr(), self.d)
+
+    def tearDown(self):
+        self.a.close()
+        self.d.close()
+        del self.a
+        del self.d
